+++ mlflow/genai/judges/utils/telemetry_utils.py
@@ -1,0 +1,69 @@
+from __future__ import annotations
+
+import logging
+
+_logger = logging.getLogger(__name__)
+
+
+def _record_judge_model_usage_success_databricks_telemetry(
+    *,
+    request_id: str | None,
+    model_provider: str,
+    endpoint_name: str,
+    num_prompt_tokens: int | None,
+    num_completion_tokens: int | None,
+) -> None:
+    try:
+        from databricks.agents.telemetry import record_judge_model_usage_success
+    except ImportError:
+        _logger.debug(
+            "Failed to import databricks.agents.telemetry.record_judge_model_usage_success; "
+            "databricks-agents needs to be installed."
+        )
+        return
+
+    from mlflow.tracking.fluent import _get_experiment_id
+    from mlflow.utils.databricks_utils import get_job_id, get_job_run_id, get_workspace_id
+
+    record_judge_model_usage_success(
+        request_id=request_id,
+        experiment_id=_get_experiment_id(),
+        job_id=get_job_id(),
+        job_run_id=get_job_run_id(),
+        workspace_id=get_workspace_id(),
+        model_provider=model_provider,
+        endpoint_name=endpoint_name,
+        num_prompt_tokens=num_prompt_tokens,
+        num_completion_tokens=num_completion_tokens,
+    )
+
+
+def _record_judge_model_usage_failure_databricks_telemetry(
+    *,
+    model_provider: str,
+    endpoint_name: str,
+    error_code: str,
+    error_message: str,
+) -> None:
+    try:
+        from databricks.agents.telemetry import record_judge_model_usage_failure
+    except ImportError:
+        _logger.debug(
+            "Failed to import databricks.agents.telemetry.record_judge_model_usage_failure; "
+            "databricks-agents needs to be installed."
+        )
+        return
+
+    from mlflow.tracking.fluent import _get_experiment_id
+    from mlflow.utils.databricks_utils import get_job_id, get_job_run_id, get_workspace_id
+
+    record_judge_model_usage_failure(
+        experiment_id=_get_experiment_id(),
+        job_id=get_job_id(),
+        job_run_id=get_job_run_id(),
+        workspace_id=get_workspace_id(),
+        model_provider=model_provider,
+        endpoint_name=endpoint_name,
+        error_code=error_code,
+        error_message=error_message,
+    )
